+++ include/ttl/nn/bits/kernels/conv.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+namespace ttl
+{
+namespace nn
+{
+namespace kernels
+{
+// template <typename D, typename R>
+// class conv1d;
+
+template <typename D, typename image_order, typename filter_order, typename R>
+class conv2d;
+}  // namespace kernels
+}  // namespace nn
+}  // namespace ttl
