+++ include/ttl/nn/kernels/cpu
@@ -3,6 +3,7 @@
 #include <ttl/nn/bits/kernels/cpu/activation.hpp>
 #include <ttl/nn/bits/kernels/cpu/bias.hpp>
 #include <ttl/nn/bits/kernels/cpu/blas.hpp>
+#include <ttl/nn/bits/kernels/cpu/conv2d.hpp>
 #include <ttl/nn/bits/kernels/cpu/elementary.hpp>
 #include <ttl/nn/bits/kernels/cpu/init.hpp>
 #include <ttl/nn/bits/kernels/cpu/loss.hpp>
